+++ KruchyPlugin1/Akcje/GenerowanieKlasyTestowej.cs
@@ -2,6 +2,7 @@
 using System.IO;
 using System.Linq;
 using System.Text;
+using System.Windows;
 using KruchyCompany.KruchyPlugin1.CodeBuilders;
 using KruchyCompany.KruchyPlugin1.Extensions;
 using KruchyCompany.KruchyPlugin1.Utils;
@@ -58,6 +59,11 @@ namespace KruchyCompany.KruchyPlugin1.Akcje
                     nazwaKlasy,
                     rodzaj,
                     interfejsTestowany);
+            if (File.Exists(pelnaSciezka))
+            {
+                MessageBox.Show("Plik już istnieje " + pelnaSciezka);
+                return;
+            }
 
             File.WriteAllText(pelnaSciezka, zawartosc, Encoding.UTF8);
             var plik = ProjektTestowy.DodajPlik(pelnaSciezka);
